fix small code issues that gcc/clang accept but MSVC correctly complains about

+++ src/visitor/MultDepthVisitor.cpp
@@ -67,7 +67,7 @@ void MultDepthVisitor::visit(Operator &elem) {
 }
 
 void MultDepthVisitor::visit(Return &elem) {
-  for (uint i = 0; i < elem.getReturnExpressions().size(); i++) {
+  for (size_t i = 0; i < elem.getReturnExpressions().size(); i++) {
     analyzeMultiplicativeDepth("<ReturnValue" + std::to_string(i) + ">", &elem, elem.getReturnExpressions().at(i));
   }
   Visitor::visit(elem);
